Update OSINT.py
Switched API decisions as an API we originally planned to use now costs money and kind of ruined the whole plan.

We are now focusing on movies/movie reviews.
I implemented an IMDB API that will allow the user to search a title and receive movies including the input as well as where they can watch the movies.
I am starting to implement a trailer API that will pull the IMDB ID from the original API and use it directly for a different API to search for a trailer