Update README.md
Added the following line:
"The tool focuses on the Internet Movie Database or (IMDB) catalog of movies. It asks for user input for keywords of a movie and returns a list of possible movies, with the correlating IMDB ID, movie trailer, where to stream the movie, cast and reviews. "

To give better insight into what the tool is designed to do.